refactor: isolate daemon remote dispatch
